Finalise essential tests

+++ peer_support/views/delete_post_view.py
@@ -4,12 +4,15 @@ from django.shortcuts import get_object_or_404
 from django.shortcuts import redirect
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib import messages
+from .helpers import get_post
 
 class DeletePostView(LoginRequiredMixin,View):
     """Delete post"""
     
     def get(self,request, post_id):
-        post = get_object_or_404(Post, pk=post_id)
+        post = get_post(request,post_id)
+        if not post:
+            return redirect('feed')
         if post.author == request.user:
             post.delete()
         else:
